+++ lsp-types/ChangeLog.md
@@ -1,5 +1,9 @@
 # Revision history for lsp-types
 
+## Unreleased
+
+- Add support for identifying client and server capabilities associated with a method.
+
 ## 2.2.0.0 -- 2024-04-29
 
 - Update the metamodel. This results in a number of breaking changes to the generated code, mostly replacing